+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#74] Add init container for dogus with volumes to execute chown on the directories with the specified uid and gid.
 
 ## [v0.22.0] - 2023-01-31
 ### Changed